// src/api/api.ts
export interface CompilationError {
    'errorCode'?: string | null;
    'message'?: string | null;
    'startLine'?: number;
    'endLine'?: number;
}
export interface CompileRequest {
    'mainFile'?: string | null;
    'optimize'?: boolean;
}
export interface CompileResult {
    'success'?: boolean;
    'output'?: string | null;
    'errors'?: Array<CompilationError> | null;
    'projectId'?: number;
    'compiledAt'?: string;
}
export interface CreateFileDto {
    'name'?: string | null;
    'path'?: string | null;
}
export interface CreateProjectRequest {
    'name'?: string | null;
}
export interface FileMetadata {
    'fileId'?: string;
    'fileName'?: string | null;
    'projectId'?: number;
    'path'?: string | null;
    'physicalPath'?: string | null;
    'existsOnDisk'?: boolean;
}
export interface PathChangeRequest {
    'oldPath'?: string | null;
    'newPath'?: string | null;
}
// ...
export interface RenameFileDto {
    'name'?: string | null;
}
export interface RenameProjectRequest {
    'newName'?: string | null;
}
export interface RunRequest {
    'mainFile'?: string | null;
    'environmentVariables'?: { [key: string]: string; } | null;
    'timeoutSeconds'?: number | null;
}
export interface RunResult {
    'success'?: boolean;
    'output'?: string | null;
    'errors'?: Array<CompilationError> | null;
    'processId'?: string | null;
    'projectId'?: number;
    'startedAt'?: string;
}
export interface RunningProjectInfo {
    'projectId'?: number;
    'processId'?: string | null;
    'startedAt'?: string;
    'uptime'?: string;
}
export interface UpdateFileDto {
    'content'?: string | null;
}
// ...
export class CompilerApi extends BaseAPI {
    /**
     * 
     * @param {number} projectId 
     * @param {CompileRequest} [compileRequest] 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public apiCompileProjectProjectIdCompilePost(projectId: number, compileRequest?: CompileRequest, options?: RawAxiosRequestConfig) {
        return CompilerApiFp(this.configuration).apiCompileProjectProjectIdCompilePost(projectId, compileRequest, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {number} projectId 
     * @param {RunRequest} [runRequest] 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public apiCompileProjectProjectIdRunPost(projectId: number, runRequest?: RunRequest, options?: RawAxiosRequestConfig) {
        return CompilerApiFp(this.configuration).apiCompileProjectProjectIdRunPost(projectId, runRequest, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {number} projectId 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public apiCompileProjectProjectIdStatusGet(projectId: number, options?: RawAxiosRequestConfig) {
        return CompilerApiFp(this.configuration).apiCompileProjectProjectIdStatusGet(projectId, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {number} projectId 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public apiCompileProjectProjectIdStopPost(projectId: number, options?: RawAxiosRequestConfig) {
        return CompilerApiFp(this.configuration).apiCompileProjectProjectIdStopPost(projectId, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public apiCompileRunningProjectsGet(options?: RawAxiosRequestConfig) {
        return CompilerApiFp(this.configuration).apiCompileRunningProjectsGet(options).then((request) => request(this.axios, this.basePath));
    }
}
// ...
export class FileApi extends BaseAPI {
    /**
     * 
     * @param {string} fileId 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public apiFilesFileIdDeletePost(fileId: string, options?: RawAxiosRequestConfig) {
        return FileApiFp(this.configuration).apiFilesFileIdDeletePost(fileId, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {string} fileId 
     * @param {string} [body] 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public apiFilesFileIdMovePost(fileId: string, body?: string, options?: RawAxiosRequestConfig) {
        return FileApiFp(this.configuration).apiFilesFileIdMovePost(fileId, body, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {string} fileId 
     * @param {RenameFileDto} [renameFileDto] 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public apiFilesFileIdRenamePost(fileId: string, renameFileDto?: RenameFileDto, options?: RawAxiosRequestConfig) {
        return FileApiFp(this.configuration).apiFilesFileIdRenamePost(fileId, renameFileDto, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {string} fileId 
     * @param {UpdateFileDto} [updateFileDto] 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public apiFilesFileIdSavePost(fileId: string, updateFileDto?: UpdateFileDto, options?: RawAxiosRequestConfig) {
        return FileApiFp(this.configuration).apiFilesFileIdSavePost(fileId, updateFileDto, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {number} projectId 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public apiFilesProjectIdGet(projectId: number, options?: RawAxiosRequestConfig) {
        return FileApiFp(this.configuration).apiFilesProjectIdGet(projectId, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {number} projectId 
     * @param {PathChangeRequest} [pathChangeRequest] 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public apiFilesProjectProjectIdChangeAllPathsPost(projectId: number, pathChangeRequest?: PathChangeRequest, options?: RawAxiosRequestConfig) {
        return FileApiFp(this.configuration).apiFilesProjectProjectIdChangeAllPathsPost(projectId, pathChangeRequest, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {number} projectId 
     * @param {CreateFileDto} [createFileDto] 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public apiFilesProjectProjectIdPost(projectId: number, createFileDto?: CreateFileDto, options?: RawAxiosRequestConfig) {
        return FileApiFp(this.configuration).apiFilesProjectProjectIdPost(projectId, createFileDto, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {string} fileId 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public apiFilesReadFileIdGet(fileId: string, options?: RawAxiosRequestConfig) {
        return FileApiFp(this.configuration).apiFilesReadFileIdGet(fileId, options).then((request) => request(this.axios, this.basePath));
    }
}